move log analysist code to snapshot_utils
